re-split services
